Modified SQLite data type mapping for bigint
Bigint doesn't exist in MySQL.  So, for those primary and foreign key
columns in ODM2 that use it, the data type has to be changed to integer.

+++ src/build_schemas/data_mapping.py
@@ -144,7 +144,7 @@ class SQLITE():
 
     def _mapSQLiteDataTypes(self, type):
         dtype_equiv = {
-            'bigint':'BIGINT',
+            'bigint':'INTEGER',
             'binary':'BINARY',
             'bit':'BIT',
             'blob':'BLOB',
